allmost done with 32k ;)

+++ src/Animable.java
@@ -3,7 +3,7 @@ public class Animable extends NodeSub {
 	public int modelHeight;
 	public VertexNormal vertexNormals[];
 
-	public void method443(int i, int j, int k, int l, int i1, int j1, int k1, int l1, int i2) {
+	public void method443(int i, int j, int k, int l, int i1, int j1, int k1, int l1, long i2) {
 		Model model = getRotatedModel();
 		if(model != null) {
 			modelHeight = model.modelHeight;
