+++ src/Object4.java
@@ -15,6 +15,6 @@ final class Object4 {
 	Animable aClass30_Sub2_Sub4_48;
 	Animable aClass30_Sub2_Sub4_49;
 	Animable aClass30_Sub2_Sub4_50;
-	int uid;
+	long uid;
 	int anInt52;
 }
